extract rules filter model to own package
