+++ pkg/security/rules/filtermodel/rule_filters_model.go
@@ -3,8 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-// Package rules holds rules related files
-package rules
+// Package filtermodel holds rules related files
+package filtermodel
 
 import (
 	"reflect"
