+++ pkg/security/rules/filtermodel/rule_filters_model_linux.go
@@ -5,8 +5,8 @@
 
 //go:build linux
 
-// Package rules holds rules related files
-package rules
+// Package filtermodel holds rules related files
+package filtermodel
 
 import (
 	"os"
